[exp10] fixed undefined behavior

+++ src/binary64/exp10/exp10.c
@@ -110,7 +110,7 @@ static inline double opolydd(double xh, double xl, int n, const double c[][2], d
 
 static inline double as_ldexp(double x, i64 i){
 #ifdef __x86_64__
-  __m128i sb; sb[0] = i<<52;
+  __m128i sb; sb[0] = (uint64_t)i<<52;
 #if defined(__clang__)
   __m128d r = _mm_set_sd(x);
 #else
